backend/requirements.txt
django
pymysql
django-cors-headers
djangorestframework
mysqlclient
